✨ feat(web): add registration endpoint in API
This commit adds a registration endpoint in the API. The `GET` method is implemented to handle the
registration process by receiving a `code` parameter from the request query string. It sends a
request to the `/register` endpoint with the provided `code` and retrieves the `token` from the
response. The `token` is then stored as a cookie with an expiration time of 30 days. Finally, a
redirect response is sent to the root URL ("/") of the application.